Remove inheritance hierarchy from Ant UI performance tests
Ant UI tests use an inheritance hierarchy for sharing configuration and
utility methods. This leads to duplications of performance test
functionality and unnecessary classes for sharing utilities.

This change removes the AbstractAntUIBuildPerformanceTest and
AbstractAntPerformanceTest classes by reusing the
PerformanceTestCaseJunit5 and integrating the remaining functionality
into the single concrete performance test classes. In order to allow
reuse of several inherited utility methods, they are moved to a
dedicated AntUITestUtil class. In addition, the functionality for
closing the welcome screen, which is embedded into the setup
functionality of the common Ant UI test superclass, is extracted into a
JUnit 5 extension to be used throughout all tests upon JUnit 5
migration.
