+++ ant/org.eclipse.ant.tests.ui/Ant Tests/org/eclipse/ant/tests/ui/separateVM/SeparateVMTests.java
@@ -14,6 +14,10 @@
 
 package org.eclipse.ant.tests.ui.separateVM;
 
+import static org.eclipse.ant.tests.ui.testplugin.AntUITestUtil.getJavaProject;
+import static org.eclipse.ant.tests.ui.testplugin.AntUITestUtil.getLaunchConfiguration;
+import static org.eclipse.ant.tests.ui.testplugin.AntUITestUtil.getProject;
+import static org.eclipse.ant.tests.ui.testplugin.AntUITestUtil.launchAndTerminate;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
@@ -32,6 +36,7 @@ import org.eclipse.ant.internal.ui.IAntUIPreferenceConstants;
 import org.eclipse.ant.launching.IAntLaunchConstants;
 import org.eclipse.ant.tests.ui.AbstractAntUIBuildTest;
 import org.eclipse.ant.tests.ui.debug.TestAgainException;
+import org.eclipse.ant.tests.ui.testplugin.AntUITestUtil;
 import org.eclipse.ant.tests.ui.testplugin.ConsoleLineTracker;
 import org.eclipse.ant.tests.ui.testplugin.ProjectHelper;
 import org.eclipse.core.resources.IFile;
@@ -306,7 +311,7 @@ public class SeparateVMTests extends AbstractAntUIBuildTest {
 		assertNotNull("Could not locate launch configuration for " + "echoingSepVM", config); //$NON-NLS-1$ //$NON-NLS-2$
 		ILaunchConfigurationWorkingCopy copy = config.getWorkingCopy();
 		copy.setAttribute(IAntUIConstants.SET_INPUTHANDLER, false);
-		launch(copy);
+		AntUITestUtil.launch(copy);
 		String message = ConsoleLineTracker.getMessage(1);
 		assertNotNull("There must be a message", message); //$NON-NLS-1$
 		assertTrue("Incorrect message. Should start with Message:. Message: " + message, message.startsWith("echo1")); //$NON-NLS-1$ //$NON-NLS-2$
