+++ ant/org.eclipse.ant.tests.ui/test plugin/org/eclipse/ant/tests/ui/testplugin/AbstractAntUITest.java
@@ -13,6 +13,9 @@
  *******************************************************************************/
 package org.eclipse.ant.tests.ui.testplugin;
 
+import static org.eclipse.ant.tests.ui.testplugin.AntUITestUtil.assertProject;
+import static org.eclipse.ant.tests.ui.testplugin.AntUITestUtil.getLaunchConfiguration;
+import static org.eclipse.ant.tests.ui.testplugin.AntUITestUtil.launchAndTerminate;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
@@ -22,37 +25,18 @@ import java.io.File;
 import java.io.IOException;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
-import java.net.URL;
 import java.nio.file.Files;
 
 import org.eclipse.ant.internal.ui.AntUIPlugin;
-import org.eclipse.ant.internal.ui.IAntUIPreferenceConstants;
 import org.eclipse.ant.internal.ui.model.AntModel;
-import org.eclipse.ant.tests.ui.debug.TestAgainException;
 import org.eclipse.ant.tests.ui.editor.support.TestLocationProvider;
 import org.eclipse.ant.tests.ui.editor.support.TestProblemRequestor;
 import org.eclipse.core.externaltools.internal.IExternalToolConstants;
 import org.eclipse.core.resources.IFile;
-import org.eclipse.core.resources.IFolder;
-import org.eclipse.core.resources.IProject;
-import org.eclipse.core.resources.ResourcesPlugin;
 import org.eclipse.core.runtime.Assert;
 import org.eclipse.core.runtime.CoreException;
-import org.eclipse.core.runtime.FileLocator;
-import org.eclipse.core.runtime.IPath;
-import org.eclipse.core.runtime.IProgressMonitor;
-import org.eclipse.core.runtime.IStatus;
-import org.eclipse.core.runtime.Status;
-import org.eclipse.core.runtime.jobs.Job;
-import org.eclipse.debug.core.DebugEvent;
-import org.eclipse.debug.core.DebugPlugin;
-import org.eclipse.debug.core.ILaunch;
 import org.eclipse.debug.core.ILaunchConfiguration;
 import org.eclipse.debug.core.ILaunchConfigurationWorkingCopy;
-import org.eclipse.debug.core.ILaunchManager;
-import org.eclipse.debug.core.model.IProcess;
-import org.eclipse.jdt.core.IJavaProject;
-import org.eclipse.jdt.core.JavaCore;
 import org.eclipse.jface.text.BadLocationException;
 import org.eclipse.jface.text.BadPositionCategoryException;
 import org.eclipse.jface.text.Document;
@@ -61,19 +45,11 @@ import org.eclipse.jface.text.IDocumentPartitioner;
 import org.eclipse.jface.text.ITypedRegion;
 import org.eclipse.jface.text.Position;
 import org.eclipse.swt.graphics.Color;
-import org.eclipse.ui.IWorkbench;
-import org.eclipse.ui.IWorkbenchWindow;
-import org.eclipse.ui.PlatformUI;
 import org.eclipse.ui.console.IHyperlink;
 import org.eclipse.ui.internal.console.ConsoleHyperlinkPosition;
 import org.eclipse.ui.internal.console.IOConsolePartition;
-import org.eclipse.ui.intro.IIntroManager;
-import org.eclipse.ui.intro.IIntroPart;
-import org.eclipse.ui.progress.UIJob;
 import org.junit.Before;
 import org.junit.Rule;
-import org.osgi.framework.Bundle;
-import org.osgi.framework.FrameworkUtil;
 
 /**
  * Abstract Ant UI test class
@@ -81,8 +57,8 @@ import org.osgi.framework.FrameworkUtil;
 @SuppressWarnings("restriction")
 public abstract class AbstractAntUITest {
 
-	public static String ANT_EDITOR_ID = "org.eclipse.ant.ui.internal.editor.AntEditor"; //$NON-NLS-1$
-	private boolean welcomeClosed = false;
+	public static final String ANT_EDITOR_ID = "org.eclipse.ant.ui.internal.editor.AntEditor"; //$NON-NLS-1$
+	private final CloseWelcomeScreenExtension closeWelcomeScreenExtension = new CloseWelcomeScreenExtension();
 	private IDocument currentDocument;
 
 	@Rule
@@ -94,7 +70,7 @@ public abstract class AbstractAntUITest {
 	 * @return the associated {@link IFile} for the given build file name
 	 */
 	protected IFile getIFile(String buildFileName) {
-		return getProject().getFolder("buildfiles").getFile(buildFileName); //$NON-NLS-1$
+		return AntUITestUtil.getProject().getFolder("buildfiles").getFile(buildFileName); //$NON-NLS-1$
 	}
 
 	/**
@@ -111,101 +87,7 @@ public abstract class AbstractAntUITest {
 	@Before
 	public void setUp() throws Exception {
 		assertProject();
-		assertWelcomeScreenClosed();
-	}
-
-	/**
-	 * Ensure the welcome screen is closed because in 4.x the debug perspective opens a giant fast-view causing issues
-	 *
-	 * @since 3.8
-	 */
-	void assertWelcomeScreenClosed() throws Exception {
-		if (!welcomeClosed && PlatformUI.isWorkbenchRunning()) {
-			final IWorkbench wb = PlatformUI.getWorkbench();
-			if (wb != null) {
-				UIJob job = new UIJob("close welcome screen for Ant test suite") { //$NON-NLS-1$
-					@Override
-					public IStatus runInUIThread(IProgressMonitor monitor) {
-						IWorkbenchWindow window = wb.getActiveWorkbenchWindow();
-						if (window != null) {
-							IIntroManager im = wb.getIntroManager();
-							IIntroPart intro = im.getIntro();
-							if (intro != null) {
-								welcomeClosed = im.closeIntro(intro);
-							}
-						}
-						return Status.OK_STATUS;
-					}
-				};
-				job.setPriority(Job.INTERACTIVE);
-				job.setSystem(true);
-				job.schedule();
-			}
-		}
-	}
-
-	/**
-	 * Asserts that the testing project has been setup in the test workspace
-	 *
-	 * @throws Exception
-	 *
-	 * @since 3.5
-	 */
-	public static void assertProject() throws Exception {
-		IProject pro = ResourcesPlugin.getWorkspace().getRoot().getProject(ProjectHelper.PROJECT_NAME);
-		if (!pro.exists()) {
-			// create project and import build files and support files
-			IProject project = ProjectHelper.createProject(ProjectHelper.PROJECT_NAME);
-			IFolder folder = ProjectHelper.addFolder(project, "buildfiles"); //$NON-NLS-1$
-			ProjectHelper.addFolder(project, "launchConfigurations"); //$NON-NLS-1$
-			File root = getFileInPlugin(ProjectHelper.TEST_BUILDFILES_DIR);
-			ProjectHelper.importFilesFromDirectory(root, folder.getFullPath(), null);
-
-			ProjectHelper.createLaunchConfigurationForBoth("echoing"); //$NON-NLS-1$
-			ProjectHelper.createLaunchConfigurationForBoth("102282"); //$NON-NLS-1$
-			ProjectHelper.createLaunchConfigurationForBoth("74840"); //$NON-NLS-1$
-			ProjectHelper.createLaunchConfigurationForBoth("failingTarget"); //$NON-NLS-1$
-			ProjectHelper.createLaunchConfiguration("build"); //$NON-NLS-1$
-			ProjectHelper.createLaunchConfiguration("bad"); //$NON-NLS-1$
-			ProjectHelper.createLaunchConfiguration("importRequiringUserProp"); //$NON-NLS-1$
-			ProjectHelper.createLaunchConfigurationForSeparateVM("echoPropertiesSepVM", "echoProperties"); //$NON-NLS-1$ //$NON-NLS-2$
-			ProjectHelper.createLaunchConfigurationForSeparateVM("extensionPointSepVM", null); //$NON-NLS-1$
-			ProjectHelper.createLaunchConfigurationForSeparateVM("extensionPointTaskSepVM", null); //$NON-NLS-1$
-			ProjectHelper.createLaunchConfigurationForSeparateVM("extensionPointTypeSepVM", null); //$NON-NLS-1$
-			ProjectHelper.createLaunchConfigurationForSeparateVM("input", null); //$NON-NLS-1$
-			ProjectHelper.createLaunchConfigurationForSeparateVM("environmentVar", null); //$NON-NLS-1$
-
-			ProjectHelper.createLaunchConfigurationForBoth("breakpoints"); //$NON-NLS-1$
-			ProjectHelper.createLaunchConfigurationForBoth("debugAntCall"); //$NON-NLS-1$
-			ProjectHelper.createLaunchConfigurationForBoth("96022"); //$NON-NLS-1$
-			ProjectHelper.createLaunchConfigurationForBoth("macrodef"); //$NON-NLS-1$
-			ProjectHelper.createLaunchConfigurationForBoth("85769"); //$NON-NLS-1$
-
-			ProjectHelper.createLaunchConfiguration("big", ProjectHelper.PROJECT_NAME + "/buildfiles/performance/build.xml"); //$NON-NLS-1$ //$NON-NLS-2$
-
-			// do not show the Ant build failed error dialog
-			AntUIPlugin.getDefault().getPreferenceStore().setValue(IAntUIPreferenceConstants.ANT_ERROR_DIALOG, false);
-		}
-	}
-
-	public static File getFileInPlugin(IPath path) {
-		try {
-			Bundle bundle = FrameworkUtil.getBundle(AbstractAntUITest.class);
-			URL installURL = bundle.getEntry("/" + path.toString()); //$NON-NLS-1$
-			URL localURL = FileLocator.toFileURL(installURL);
-			return new File(localURL.getFile());
-		} catch (IOException e) {
-			return null;
-		}
-	}
-
-	/**
-	 * Returns the 'AntUITests' project.
-	 *
-	 * @return the test project
-	 */
-	protected static IProject getProject() {
-		return ResourcesPlugin.getWorkspace().getRoot().getProject(ProjectHelper.PROJECT_NAME);
+		closeWelcomeScreenExtension.assertWelcomeScreenClosed();
 	}
 
 	/**
@@ -301,82 +183,6 @@ public abstract class AbstractAntUITest {
 		ILaunchConfigurationWorkingCopy copy = config.getWorkingCopy();
 		copy.setAttribute(IExternalToolConstants.ATTR_TOOL_ARGUMENTS, arguments);
 		launchAndTerminate(copy, 20000);
-	}
-
-	/**
-	 * Returns the launch configuration for the given build file
-	 *
-	 * @param buildFileName
-	 *            build file to launch
-	 */
-	protected ILaunchConfiguration getLaunchConfiguration(String buildFileName) {
-		IFile file = getJavaProject().getProject().getFolder("launchConfigurations").getFile(buildFileName + ".launch"); //$NON-NLS-1$ //$NON-NLS-2$
-		ILaunchConfiguration config = getLaunchManager().getLaunchConfiguration(file);
-		assertTrue("Could not find launch configuration for " + buildFileName, config.exists()); //$NON-NLS-1$
-		return config;
-	}
-
-	/**
-	 * Returns the launch manager
-	 *
-	 * @return launch manager
-	 */
-	public static ILaunchManager getLaunchManager() {
-		return DebugPlugin.getDefault().getLaunchManager();
-	}
-
-	/**
-	 * Returns the 'AntUITests' project.
-	 *
-	 * @return the test project
-	 */
-	public static IJavaProject getJavaProject() {
-		return JavaCore.create(getProject());
-	}
-
-	/**
-	 * Launches the given configuration and waits for the terminated event or the length of the given timeout, whichever comes first
-	 */
-	protected void launchAndTerminate(ILaunchConfiguration config, int timeout) throws CoreException {
-		DebugEventWaiter waiter = new DebugElementKindEventWaiter(DebugEvent.TERMINATE, IProcess.class);
-		waiter.setTimeout(timeout);
-
-		Object terminatee = launchAndWait(config, waiter);
-		assertTrue("terminatee is not an IProcess", terminatee instanceof IProcess); //$NON-NLS-1$
-		IProcess process = (IProcess) terminatee;
-		boolean terminated = process.isTerminated();
-		assertTrue("process is not terminated", terminated); //$NON-NLS-1$
-	}
-
-	/**
-	 * Launches the given configuration and waits for an event. Returns the source of the event. If the event is not received, the launch is
-	 * terminated and an exception is thrown.
-	 *
-	 * @param configuration
-	 *            the configuration to launch
-	 * @param waiter
-	 *            the event waiter to use
-	 * @return Object the source of the event
-	 */
-	protected Object launchAndWait(ILaunchConfiguration configuration, DebugEventWaiter waiter) throws CoreException {
-		ILaunch launch = configuration.launch(ILaunchManager.RUN_MODE, null);
-		Object suspendee = waiter.waitForEvent();
-		if (suspendee == null) {
-			try {
-				launch.terminate();
-			}
-			catch (CoreException e) {
-				e.printStackTrace();
-			}
-			throw new TestAgainException("Retest - Program did not suspend launching: " + configuration.getName()); //$NON-NLS-1$
-		}
-		boolean terminated = launch.isTerminated();
-		assertTrue("launch did not terminate", terminated); //$NON-NLS-1$
-		if (terminated && !ConsoleLineTracker.isClosed()) {
-			ConsoleLineTracker.waitForConsole();
-		}
-		assertTrue("Console is not closed", ConsoleLineTracker.isClosed()); //$NON-NLS-1$
-		return suspendee;
 	}
 
 	/**
